Fix bug with inotify event handling
The monitoring of local directories uses inotify on Linux. For handling
of recursive directory monitoring, an extra event 'CREATE' is monitored.
Up until now this also caused the CREATE event to be handled for each
file. This in turn often caused handling of multiple events for the same
file. This fix checks if the event matches the configured events.